refactor: move user form to separate component
